mern 08 authentication done

+++ mern_stack_ninja08/frontend/src/components/WorkoutDetails.js
@@ -1,12 +1,19 @@
 import { useWorkoutContext } from "../hooks/useWorkoutsContext";
+import { useAuthContext } from "../hooks/useAuthContext";
 import formatDistanceToNow from "date-fns/formatDistanceToNow";
 
 function WorkoutDetails({ workout }) {
     const { dispatch } = useWorkoutContext();
+    const { user } = useAuthContext();
 
     const handleDelete = async () => {
+        if (!user) return;
+
         const response = await fetch("/api/workouts/" + workout._id, {
             method: "DELETE",
+            headers: {
+                authorization: `Bearer ${user.token}`,
+            },
         });
 
         if (response.ok) {
